Add "all" tabs on pokemons and synergies wiki
