+++ app/public/src/pages/component/typeahead/pokemon-typeahead.css
@@ -1,10 +1,11 @@
 .pokemon-typeahead .form-control {
   color: #fff;
-  background-color: rgb(84, 89, 107);
+  background-color: #40404080;
   font-size: 1em;
   border-radius: 12px;
   width: 100%;
   border: 4px solid black;
+  padding: 0 0.75rem;
 }
 
 .pokemon-typeahead .form-control::placeholder {
