+++ app/public/src/pages/component/wiki/wiki.css
@@ -106,6 +106,20 @@
   padding: 0;
 }
 
+#wiki-pokemons-all tr {
+  border-top: 4px solid currentColor;
+}
+
+#wiki-types-all tr {
+  border-top: 4px solid #40404080;
+}
+
+#wiki-pokemons-all td:nth-child(2),
+#wiki-types-all td:nth-child(2) {
+  display: flex;
+  flex-wrap: wrap;
+}
+
 #wiki-weather > ul {
   display: grid;
   grid-template-columns: repeat(auto-fit, minmax(20em, 1fr));
